+++ src/lib/packages/calendar/core/src/cell/index.ts
@@ -1,2 +1,2 @@
-export * from './Cell'
-export * from './CellList'
+export * from './Cell.js'
+export * from './CellList.js'