+++ src/lib/packages/calendar/core/src/disabled/DisabledList.ts
@@ -1,5 +1,5 @@
-import type{ Cell } from "../cell";
-import { Disabled } from "./Disabled";
+import type{ Cell } from "../cell/index.js";
+import { Disabled } from "./Disabled.js";
 
 export class DisabledList<T extends Cell[]> {
     list