+++ src/lib/packages/forms/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @cluue/forms
 
+## 1.0.20
+
+### Patch Changes
+
+- Upgrade perfomance
+- Updated dependencies
+  - @cluue/popover@1.0.15
+  - @cluue/base@1.0.20
+
 ## 1.0.19
 
 ### Patch Changes